lots, i'm bad

+++ lib/athenry/setup.rb
@@ -1,28 +1,25 @@
 module Athenry
   class Setup < Helper
+    
     def fetch
-      announcing "Fetching #{CONFIG['WORKDIR']}" do
-        silent "wget -c #{CONFIG['STAGEURL']} -O #{CONFIG['WORKDIR']}/stage3-amd64-current.tar.bz2"
+      announcing "Fetching" do
+        unless File.exists?("#{CONFIG['WORKDIR']}/stage3-amd64-current.tar.bz2")
+          silent "wget -c #{CONFIG['STAGEURL']} -O #{CONFIG['WORKDIR']}/stage3-amd64-current.tar.bz2"
+        end
       end
     end
 
     def extract
       announcing "Extracting stage file" do
-        silent  "sudo tar xjpf stage3-amd64-current.tar.bz2 -C #{CONFIG['BASE']}"
-      end
-    end
-
-    def mount
-      announcing "Mounting dev, sys, and proc" do
-        silent "sudo mount -o bind /dev #{CONFIG['BASE']}/dev"
-        silent "sudo mount -o bind /sys #{CONFIG['BASE']}"
-        silent "sudo mount -t proc none #{CONFIG['BASE']}/proc"
+        silent "sudo tar xjpf stage3-amd64-current.tar.bz2 -C #{CONFIG['BASE']}"
       end
     end
 
     def snapshot
-      announcing "Fetching portage snapshot" do
-        silent "wget -c #{CONFIG['SNAPSHOTURL']} -O #{CONFIG['WORK']}/portage-latest.tar.bz2"
+      announcing "Fetching and extracting portage snapshot" do
+        unless File.exists?("#{CONFIG['WORKDIR']}/portage-latest.tar.bz2")
+          silent "wget -c #{CONFIG['SNAPSHOTURL']} -O #{CONFIG['WORK']}/portage-latest.tar.bz2"
+        end
         silent "sudo tar xjpf #{CONFIG['WORK']}/portage-latest.tar.bz2 -C #{CONFIG['BASE']}/usr"
       end
     end
